Modifications and hacks... aint pretty
